+++ system/drivers/DC_File.php
@@ -335,7 +335,7 @@ window.addEvent(\'domready\', function() {
 			if ($this->Input->post('saveNclose'))
 			{
 				$this->resetMessages();
-				setcookie('BE_PAGE_OFFSET', 0, 0, '/');
+				$this->setCookie('BE_PAGE_OFFSET', 0, 0);
 				$this->redirect($this->getReferer());
 			}
 
@@ -427,7 +427,7 @@ window.addEvent(\'domready\', function() {
 		}
 
 		// Save the value if there was no error
-		if ((strlen($varValue) || !$arrData['eval']['doNotSaveEmpty']) && $strCurrent != $varValue)
+		if ((strlen($varValue) || !$arrData['eval']['doNotSaveEmpty']) && $strCurrent !== $varValue)
 		{
 			$strKey = sprintf("\$GLOBALS['TL_CONFIG']['%s']", $this->strField);
 			$this->Config->update($strKey, $varValue);