+++ system/modules/backend/StyleSheets.php
@@ -1177,7 +1177,7 @@ class StyleSheets extends Backend
 			}
 
 			// Redirect
-			setcookie('BE_PAGE_OFFSET', 0, 0, '/');
+			$this->setCookie('BE_PAGE_OFFSET', 0, 0);
 			$this->redirect(str_replace('&key=import', '', $this->Environment->request));
 		}