+++ system/modules/frontend/FrontendTemplate.php
@@ -129,10 +129,16 @@ class FrontendTemplate extends Template
 			}
 		}
 
-		$intCache = null;
+		$intCache = 0;
 
-		// Cache the page if it is not protected
-		if (empty($_POST) && !BE_USER_LOGGED_IN && !FE_USER_LOGGED_IN && !$_SESSION['DISABLE_CACHE'] && !isset($_SESSION['LOGIN_ERROR']) && ($GLOBALS['TL_CONFIG']['cacheMode'] == 'both' || $GLOBALS['TL_CONFIG']['cacheMode'] == 'server') && intval($objPage->cache) > 0 && !$objPage->protected)
+		// Decide whether the page shall be cached
+		if (empty($_POST) && !BE_USER_LOGGED_IN && !FE_USER_LOGGED_IN && !$_SESSION['DISABLE_CACHE'] && !isset($_SESSION['LOGIN_ERROR']) && intval($objPage->cache) > 0 && !$objPage->protected)
+		{
+			$intCache = time() + intval($objPage->cache);
+		}
+
+		// Server-side cache
+		if ($intCache > 0 && ($GLOBALS['TL_CONFIG']['cacheMode'] == 'both' || $GLOBALS['TL_CONFIG']['cacheMode'] == 'server'))
 		{
 			// If the request string is empty, use a special cache tag which considers the page language
 			if ($this->Environment->request == '' || $this->Environment->request == 'index.php')
@@ -156,7 +162,6 @@ class FrontendTemplate extends Template
 
 			// Replace insert tags for caching
 			$strBuffer = $this->replaceInsertTags($strBuffer, true);
-			$intCache = intval($objPage->cache) + time();
 
 			// Create the cache file
 			$objFile = new File('system/tmp/' . md5($strCacheKey) . '.html');
@@ -165,12 +170,12 @@ class FrontendTemplate extends Template
 			$objFile->close();
 		}
 
-		// Send cache headers
+		// Client-side cache
 		if (!headers_sent())
 		{
-			if ($intCache !== null && ($GLOBALS['TL_CONFIG']['cacheMode'] == 'both' || $GLOBALS['TL_CONFIG']['cacheMode'] == 'browser'))
+			if ($intCache > 0 && ($GLOBALS['TL_CONFIG']['cacheMode'] == 'both' || $GLOBALS['TL_CONFIG']['cacheMode'] == 'browser'))
 			{
-				header('Cache-Control: public, max-age=' . ($intCache -  time()));
+				header('Cache-Control: public, max-age=' . ($intCache - time()));
 				header('Expires: ' . gmdate('D, d M Y H:i:s', $intCache) . ' GMT');
 				header('Last-Modified: ' . gmdate('D, d M Y H:i:s', time()) . ' GMT');
 				header('Pragma: public');